+++ app/src/main/java/com/example/water_sentinel/DetailsActivity.kt
@@ -25,6 +25,7 @@ import android.view.MotionEvent
 import android.view.View
 import android.widget.TextView
 import androidx.appcompat.widget.Toolbar
+import com.example.water_sentinel.ui.dashboard.DashboardActivity
 
 class DetailsActivity : AppCompatActivity() {
