+++ app/src/main/java/com/example/water_sentinel/data/db/DataHistory.kt
@@ -1,4 +1,4 @@
-package com.example.water_sentinel.db
+package com.example.water_sentinel.data.db
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
