+++ app/src/main/java/com/example/water_sentinel/data/db/TodoDao.kt
@@ -1,4 +1,4 @@
-package com.example.water_sentinel.db
+package com.example.water_sentinel.data.db
 
 import androidx.room.Dao
 import androidx.room.Insert
