+++ app/src/main/java/com/example/water_sentinel/data/repository/DataRepository.kt
@@ -1,0 +1,4 @@
+package com.example.water_sentinel.data.repository
+
+class DataRepository {
+}
