+++ app/src/main/java/com/example/water_sentinel/service/DataCollectionService.kt
@@ -1,0 +1,4 @@
+package com.example.water_sentinel.service
+
+class DataCollectionService {
+}
