+++ app/src/main/java/com/example/water_sentinel/ui/splash/SplashActivity.kt
@@ -1,14 +1,12 @@
-package com.example.water_sentinel
+package com.example.water_sentinel.ui.splash
 
 import android.content.Intent
-//import android.graphics.Color
 import android.os.Bundle
 import android.os.Handler
 import android.os.Looper
 import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
-//import androidx.core.view.ViewCompat
-//import androidx.core.view.WindowInsetsCompat
+import com.example.water_sentinel.ui.welcome.WelcomeActivity
 
 class SplashActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
