+++ app/src/main/java/com/example/water_sentinel/util/PermissionHelper.kt
@@ -1,0 +1,4 @@
+package com.example.water_sentinel.util
+
+class PermissionHelper {
+}
